Expand documentation of `git` dependency subkeys
I need to pin a dependency to a specific git commit hash. After reading
these docs, it was not clear to me that I could do so with the `rev`
key, because "something else" is vague. I wound up digging into the
source code and learned what I needed there.

This change expands the documentation for `git` dependencies to explain
the sub-keys in more detail and give examples of how to use each.