repo: handle invalid distributions gracefully
This change ensures that poetry does not bail out when encountering a
bad distributions in `sys_path`. This behaviour is similar to how `pip`
handles this today. In addition to ignoring these distributions, we
also issue a warning log so users can choose to act on this.

Further, this change also handles a scenario where an empty path is
present in the `sys_path`.